requirements.txt
fastapi>=0.110
uvicorn>=0.25
pytest>=7.0
httpx>=0.24
ollama>=0.3
discord.py>=2.3
python-dotenv>=1.0
